requirements.txt
Package                                  Version
---------------------------------------- ------------
aiohappyeyeballs                         2.6.1
aiohttp                                  3.13.5
aiosignal                                1.4.0
altair                                   6.0.0
annotated-doc                            0.0.4
annotated-types                          0.7.0
anthropic                                0.89.0
anyio                                    4.13.0
attrs                                    26.1.0
bcrypt                                   5.0.0
blinker                                  1.9.0
build                                    1.4.2
cachetools                               7.0.5
certifi                                  2026.2.25
charset-normalizer                       3.4.7
chromadb                                 1.5.5
click                                    8.3.2
dataclasses-json                         0.6.7
distro                                   1.9.0
docstring-parser                         0.17.0
durationpy                               0.10
fastapi                                  0.104.1
filelock                                 3.25.2
flatbuffers                              25.12.19
frozenlist                               1.8.0
fsspec                                   2026.3.0
genai-prices                             0.0.56
gitdb                                    4.0.12
gitpython                                3.1.46
googleapis-common-protos                 1.74.0
griffelib                                2.0.2
grpcio                                   1.80.0
h11                                      0.16.0
hf-xet                                   1.4.3
httpcore                                 1.0.9
httptools                                0.7.1
httpx                                    0.28.1
httpx-sse                                0.4.3
huggingface-hub                          1.9.0
idna                                     3.11
importlib-metadata                       8.7.1
importlib-resources                      6.5.2
jinja2                                   3.1.6
jiter                                    0.13.0
joblib                                   1.5.3
jsonpatch                                1.33
jsonpointer                              3.1.1
jsonschema                               4.26.0
jsonschema-specifications                2025.9.1
kubernetes                               35.0.0
langchain-classic                        1.0.3
langchain-community                      0.4.1
langchain-core                           1.2.26
langchain-text-splitters                 1.1.1
langsmith                                0.7.26
logfire-api                              4.31.0
markdown-it-py                           4.0.0
markupsafe                               3.0.3
marshmallow                              3.26.2
mdurl                                    0.1.2
mmh3                                     5.2.1
mpmath                                   1.3.0
multidict                                6.7.1
mypy-extensions                          1.1.0
narwhals                                 2.19.0
networkx                                 3.6.1
numpy                                    2.4.4
oauthlib                                 3.3.1
onnxruntime                              1.24.4
opentelemetry-api                        1.40.0
opentelemetry-exporter-otlp-proto-common 1.40.0
opentelemetry-exporter-otlp-proto-grpc   1.40.0
opentelemetry-proto                      1.40.0
opentelemetry-sdk                        1.40.0
opentelemetry-semantic-conventions       0.61b0
orjson                                   3.11.8
overrides                                7.7.0
packaging                                26.0
pandas                                   3.0.2
pillow                                   12.2.0
propcache                                0.4.1
protobuf                                 6.33.6
pyarrow                                  23.0.1
pybase64                                 1.4.3
pydantic                                 2.12.5
pydantic-ai-slim                         1.77.0
pydantic-core                            2.41.5
pydantic-evals                           1.77.0
pydantic-graph                           1.77.0
pydantic-settings                        2.13.1
pydeck                                   0.9.1
pygments                                 2.20.0
pymupdf                                  1.27.2.2
pypika                                   0.51.1
pyproject-hooks                          1.2.0
python-dateutil                          2.9.0.post0
python-dotenv                            1.2.2
python-multipart                         0.0.6
pytz                                     2026.1.post1
pyyaml                                   6.0.3
referencing                              0.37.0
regex                                    2026.4.4
requests                                 2.33.1
requests-oauthlib                        2.0.0
requests-toolbelt                        1.0.0
rich                                     14.3.3
rpds-py                                  0.30.0
safetensors                              0.7.0
scikit-learn                             1.8.0
scipy                                    1.17.1
sentence-transformers                    5.3.0
setuptools                               81.0.0
shellingham                              1.5.4
six                                      1.17.0
smmap                                    5.0.3
sniffio                                  1.3.1
sqlalchemy                               2.0.49
starlette                                0.27.0
streamlit                                1.56.0
sympy                                    1.14.0
tenacity                                 9.1.4
threadpoolctl                            3.6.0
tokenizers                               0.22.2
toml                                     0.10.2
torch                                    2.11.0
tornado                                  6.5.5
tqdm                                     4.67.3
transformers                             5.5.0
typer                                    0.24.1
typing-extensions                        4.15.0
typing-inspect                           0.9.0
typing-inspection                        0.4.2
tzdata                                   2026.1
tzlocal                                  5.3.1
urllib3                                  2.6.3
uuid-utils                               0.14.1
uvicorn                                  0.44.0
uvloop                                   0.22.1
validators                               0.35.0
watchfiles                               1.1.1
websocket-client                         1.9.0
websockets                               16.0
xxhash                                   3.6.0
yarl                                     1.23.0
zipp                                     3.23.0
zstandard                                0.25.0
